add transaction with resolving payment status using by FakePaymentGatewayService

+++ app/Actions/Checkout/ConfirmPaymentAction.php
@@ -2,11 +2,14 @@
 
 namespace App\Actions\Checkout;
 
+use App\Data\ConfirmPaymentData;
 use App\Enums\OrderStatus;
+use App\Enums\PaymentStatus;
 use App\Exceptions\ProductVariantOutOfStockException;
-use App\Models\Order\Order;
+use App\Models\Payment\Payment;
 use App\Repositories\OrderRepository;
 use App\Services\Cart\CurrentCartService;
+use App\Services\FakePaymentGatewayService;
 use Illuminate\Support\Facades\DB;
 use Throwable;
 
@@ -14,36 +17,69 @@ readonly class ConfirmPaymentAction
 {
     public function __construct(
         private OrderRepository $orderRepository,
-        private CurrentCartService $currentCartService
+        private CurrentCartService $currentCartService,
+        private FakePaymentGatewayService $fakePaymentGateway
     ) {
     }
 
     /**
      * @throws Throwable
      */
-    public function handle(int $orderId): Order
+    public function handle(ConfirmPaymentData $data, int $orderId): array
     {
-        return DB::transaction(function () use ($orderId): Order  {
+        return DB::transaction(function () use ($data, $orderId): array  {
             $order = $this->orderRepository->findById($orderId);
 
-            foreach ($order->items as $orderItem) {
-                $productVariant = $orderItem->productVariant()->lockForUpdate()->first();
-
-                if ($productVariant->stock < $orderItem->quantity) {
-                    throw new ProductVariantOutOfStockException($productVariant->id);
-                }
-
-                $productVariant->decrement('stock', $orderItem->quantity);
-            }
-
             $order->update([
-                'status' => OrderStatus::PAID
+                'status' => OrderStatus::PENDING
             ]);
 
-            $cart = $this->currentCartService->findById();
-            $cart->items()->delete();
+            $chance = $this->fakePaymentGateway->emulation();
 
-            return $order;
+            if ($chance) {
+                foreach ($order->items as $orderItem) {
+                    $productVariant = $orderItem->productVariant()->lockForUpdate()->first();
+
+                    if ($productVariant->stock < $orderItem->quantity) {
+                        throw new ProductVariantOutOfStockException($productVariant->id);
+                    }
+
+                    $productVariant->decrement('stock', $orderItem->quantity);
+                }
+
+                $payment = Payment::create([
+                    'order_id' => $order->id,
+                    'amount' => $order->total,
+                    'status' => PaymentStatus::SUCCESS,
+                    'card_last4' => substr($data->getLastCard4(), -4),
+                    'gateway' => 'Monobank'
+                ]);
+
+                $order->update([
+                    'status' => OrderStatus::PAID
+                ]);
+
+                $cart = $this->currentCartService->findById();
+                $cart->items()->delete();
+            } else {
+                $payment = Payment::create([
+                    'order_id' => $order->id,
+                    'amount' => $order->total,
+                    'status' => PaymentStatus::FAILED,
+                    'card_last4' => substr($data->getLastCard4(), -4),
+                    'gateway' => 'Monobank'
+                ]);
+
+                $order->update([
+                    'status' => OrderStatus::CANCELLED
+                ]);
+            }
+
+            return [
+                'order' => $order,
+                'payment' => $payment,
+                'chance' => $chance
+            ];
         });
     }
 }
